Move Ingestion to usecase to avoid circular dependency
